Create a robust autoencoder

A robust autoencoder uses a special objective function, correntropy, a localized similarity measure which makes it less sensitive to noise in data. Correntropy specifically measures the probability density that two events are equal, and is less affected by outliers than the mean squared error.

autoencoder_robust(network, sigma = 0.2)

Arguments

network
Layer construct of class "ruta_network"

sigma
Sigma parameter in the kernel used for correntropy

Value

A construct of class "ruta_autoencoder"

References

Robust feature learning by stacked autoencoder with maximum correntropy criterion

See also