Add denoising behavior to any autoencoder
Source: R/autoencoder_denoising.R

Converts an autoencoder into a denoising one by adding a filter for the input data

make_denoising(learner, ratio = 0.05)

Arguments

learner
The "ruta_autoencoder" object

ratio
The ratio of each instance in the input data which will be altered by random noise

Value

An autoencoder object which contains the noisy filter