Save and load Ruta models

Functions to save a trained or untrained Ruta learner into a file and load it

save_as(learner, file = paste0(substitute(learner), ".tar.gz"),
  compression = "gzip")

load_from(file)

Arguments

learner
The "ruta_autoencoder" object to be saved

file
File name with extension (usually .tar.gz) where the object will be saved to or loaded from

compression
Type of compression to be used, for R function tar

Value

save_as invisibly returns the filename where the model has been saved, load_from returns the loaded model as a "ruta_autoencoder" object

Examples

# NOT RUN {
x <- as.matrix(iris[, 1:4])

# Save a trained model
autoencoder(2) %>% train(x) %>% save_as("my_model.tar.gz")

# Load and use the model
encoded <- load_from("my_model.tar.gz") %>% encode(x)
# }