Obtain a Keras block of layers for the variational autoencoder

This block contains two dense layers representing the mean and log var of a Gaussian distribution and a sampling layer.

# S3 method for ruta_layer_variational
to_keras(x, input_shape, model = keras::keras_model_sequential(), ...)

Arguments

x

The layer object

input_shape

Number of features in training data

model

Keras model where the layers will be added

...

Unused

Value

A Layer object from Keras

References

Auto-Encoding Variational Bayes

Under the Hood of the Variational Autoencoder (in Prose and Code)

Keras example: Variational autoencoder