Translate sparsity regularization to Keras regularizer

Translate sparsity regularization to Keras regularizer

# S3 method for ruta_sparsity
to_keras(x, activation)

Arguments

x
Sparsity object

activation
Name of the activation function used in the encoding layer

Value

Function which can be used as activity regularizer in a Keras layer

References

Sparse deep belief net model for visual area V2

Andrew Ng, Sparse Autoencoder. CS294A Lecture Notes (2011)