Create a denoising autoencoder

A denoising autoencoder trains with noisy data in order to create a model able to reduce noise in reconstructions from input data

autoencoder_denoising(network, loss = "mean_squared_error",
  noise_type = "zeros", ...)

Arguments

network
Layer construct of class "ruta_network"

loss
Loss function to be optimized

noise_type
Type of data corruption which will be used to train the autoencoder, as a character string. Available types:

"zeros" Randomly set components to zero (noise_zeros)

"ones" Randomly set components to one (noise_ones)

"saltpepper" Randomly set components to zero or one (noise_saltpepper)

"gaussian" Randomly offset each component of an input as drawn from Gaussian distributions with the same variance (additive Gaussian noise, noise_gaussian)

"cauchy" Randomly offset each component of an input as drawn from Cauchy distributions with the same scale (additive Cauchy noise, noise_cauchy)

...
Extra parameters to customize the noisy filter:

p The probability that each instance in the input data which will be altered by random noise (for "zeros", "ones" and "saltpepper")

var or sd The variance or standard deviation of the Gaussian distribution from which additive noise will be drawn (for "gaussian", only one of those parameters is necessary)

scale For the Cauchy distribution

Value

A construct of class "ruta_autoencoder"

References

Extracting and composing robust features with denoising autoencoders

See also