Train a learner object with data

This function compiles the neural network described by the learner object and trains it with the input data.

train(learner, ...)

# S3 method for ruta_autoencoder
train(learner, data, validation_data = NULL,
  epochs = 100, ...)

Arguments

learner
A "ruta_autoencoder" object

...
Additional parameters for keras::fit

data
Training data: columns are attributes and rows are instances

validation_data
Additional data.frame of data which will not be used for training but the loss measure will be calculated against it

epochs
The number of times data will pass through the network

Value

Same autoencoder passed as parameter, with trained internal models